Move unit test utility functions to unit folder
