+++ src/geo/transform.test.ts
@@ -2,7 +2,7 @@ import Point from '@mapbox/point-geometry';
 import Transform from './transform';
 import LngLat from './lng_lat';
 import {OverscaledTileID, CanonicalTileID} from '../source/tile_id';
-import {fixedLngLat, fixedCoord} from '../../test/util/fixed';
+import {fixedLngLat, fixedCoord} from '../../test/unit/util/fixed';
 
 describe('transform', () => {
     test('creates a transform', () => {
