+++ src/ui/handler/box_zoom.test.ts
@@ -1,6 +1,6 @@
 import Map from '../map';
 import DOM from '../../util/dom';
-import simulate from '../../../test/util/simulate_interaction';
+import simulate from '../../../test/unit/util/simulate_interaction';
 import {setWebGlContext, setPerformance, setMatchMedia} from '../../util/test/util';
 
 function createMap(clickTolerance) {
