+++ src/ui/handler/keyboard.test.ts
@@ -1,6 +1,6 @@
 import Map from '../../ui/map';
 import DOM from '../../util/dom';
-import simulate from '../../../test/util/simulate_interaction';
+import simulate from '../../../test/unit/util/simulate_interaction';
 import {extend} from '../../util/util';
 import {setMatchMedia, setPerformance, setWebGlContext} from '../../util/test/util';
