+++ src/ui/handler/map_event.test.ts
@@ -1,6 +1,6 @@
 import Map, {MapOptions} from '../map';
 import DOM from '../../util/dom';
-import simulate from '../../../test/util/simulate_interaction';
+import simulate from '../../../test/unit/util/simulate_interaction';
 import {setMatchMedia, setPerformance, setWebGlContext} from '../../util/test/util';
 
 function createMap() {
