+++ src/ui/handler/scroll_zoom.test.ts
@@ -1,7 +1,7 @@
 import browser from '../../util/browser';
 import Map from '../../ui/map';
 import DOM from '../../util/dom';
-import simulate from '../../../test/util/simulate_interaction';
+import simulate from '../../../test/unit/util/simulate_interaction';
 import {setMatchMedia, setPerformance, setWebGlContext} from '../../util/test/util';
 
 function createMap() {
