+++ src/ui/map.test.ts
@@ -4,8 +4,8 @@ import LngLat from '../geo/lng_lat';
 import Tile from '../source/tile';
 import {OverscaledTileID} from '../source/tile_id';
 import {Event, ErrorEvent} from '../util/evented';
-import simulate from '../../test/util/simulate_interaction';
-import {fixedLngLat, fixedNum} from '../../test/util/fixed';
+import simulate from '../../test/unit/util/simulate_interaction';
+import {fixedLngLat, fixedNum} from '../../test/unit/util/fixed';
 import {LayerSpecification, SourceSpecification, StyleSpecification} from '../style-spec/types';
 import {RequestTransformFunction} from '../util/request_manager';
 import {extend} from '../util/util';
