+++ src/ui/map/isRotating.test.ts
@@ -1,6 +1,6 @@
 import Map from '../../ui/map';
 import DOM from '../../util/dom';
-import simulate from '../../../test/util/simulate_interaction';
+import simulate from '../../../test/unit/util/simulate_interaction';
 import browser from '../../util/browser';
 import {setWebGlContext, setPerformance, setMatchMedia} from '../../util/test/util';
