+++ src/ui/popup.test.ts
@@ -2,7 +2,7 @@ import {createMap as globalCreateMap, setPerformance, setWebGlContext} from '../
 import Popup, {Offset} from './popup';
 import LngLat from '../geo/lng_lat';
 import Point from '@mapbox/point-geometry';
-import simulate from '../../test/util/simulate_interaction';
+import simulate from '../../test/unit/util/simulate_interaction';
 import {PositionAnchor} from './anchor';
 
 const containerWidth = 512;
