+++ test/unit/util/simulate_interaction.ts
